# campus-recruitment-plan/校园招聘接口文档包/OpenAPI接口草案.yaml
openapi: 3.1.0
info:
  title: 校园智能招聘与面试预约平台 API
  version: 1.0.0
  description: 面向学生、企业 HR、管理员的校园招聘与面试预约平台 OpenAPI 草案。
servers:
- url: http://localhost:8080/api
  description: 本地开发环境
- url: https://api.example.com/api
  description: 生产环境占位
...
security:
- bearerAuth: []
paths:
...
  /applications:
    post:
      tags:
      - Application
      summary: 投递岗位
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseApplication'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreateRequest'
  /applications/my:
    get:
      tags:
      - Application
      summary: 我的投递列表
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseApplicationPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        required: false
        description: 投递状态
        schema:
          type: string
          enum:
          - DELIVERED
          - VIEWED
          - INTERVIEW_INVITED
          - BOOKED
          - REJECTED
          - CANCELED
          - FINISHED
      - *id001
      - *id002
...
  /admin/dashboard:
    get:
      tags:
      - Admin
      summary: 管理后台数据看板
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminDashboard'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /admin/users:
    get:
      tags:
      - Admin
      summary: 用户列表
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminUserPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: userType
        in: query
        required: false
        description: 用户类型
        schema:
          type: string
          enum:
          - STUDENT
          - COMPANY
          - ADMIN
      - name: status
        in: query
        required: false
        description: 状态
        schema:
          type: string
      - name: keyword
        in: query
        required: false
        description: 用户名/手机号/邮箱
        schema:
          type: string
      - *id001
      - *id002
  /admin/users/{id}/disable:
    put:
      tags:
      - Admin
      summary: 禁用用户
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 用户 ID
        schema:
          type: integer
          format: int64
  /admin/users/{id}/enable:
    put:
      tags:
      - Admin
      summary: 恢复用户
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 用户 ID
        schema:
          type: integer
          format: int64
...
  /admin/companies/{id}/audit:
    put:
      tags:
      - Admin
      summary: 审核企业认证
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 企业资料 ID
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditRequest'
...
  /admin/jobs/{id}/audit:
    put:
      tags:
      - Admin
      summary: 审核岗位
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 岗位 ID
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditRequest'
...
  /admin/logs/operation:
    get:
      tags:
      - Log
      summary: 操作日志列表
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOperationLogPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: username
        in: query
        required: false
        description: 用户名
        schema:
          type: string
      - name: module
        in: query
        required: false
        description: 模块
        schema:
          type: string
      - name: resultStatus
        in: query
        required: false
        description: 结果状态
        schema:
          type: string
      - *id001
      - *id002
  /admin/logs/login:
    get:
      tags:
      - Log
      summary: 登录日志列表
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginLogPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: username
        in: query
        required: false
        description: 用户名
        schema:
          type: string
      - name: status
        in: query
        required: false
        description: 登录状态
        schema:
          type: string
      - *id001
      - *id002
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserType:
      type: string
      enum:
      - STUDENT
      - COMPANY
      - ADMIN
    UserStatus:
      type: string
      enum:
      - NORMAL
      - DISABLED
      - PENDING
...
    ApplicationStatus:
      type: string
      enum:
      - DELIVERED
      - VIEWED
      - INTERVIEW_INVITED
      - BOOKED
      - REJECTED
      - CANCELED
      - FINISHED
...
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 30002
        message:
          type: string
          example: 重复投递
        data:
          nullable: true
...
    ApplicationCreateRequest:
      type: object
      required:
      - jobId
      properties:
        jobId:
          type: integer
          format: int64
        resumeId:
          type: integer
          format: int64
          description: 不传则使用默认简历
...
    Application:
      type: object
      properties:
        id:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        studentName:
          type: string
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        jobId:
          type: integer
          format: int64
        jobTitle:
          type: string
        resumeId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        applyTime:
          type: string
          format: date-time
...
    AdminDashboard:
      type: object
      properties:
        userCount:
          type: integer
        companyCount:
          type: integer
        jobCount:
          type: integer
        todayApplicationCount:
          type: integer
        pendingCompanyAuditCount:
          type: integer
        pendingJobAuditCount:
          type: integer
    AdminUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        nickname:
          type: string
        phone:
          type: string
        email:
          type: string
        userType:
          $ref: '#/components/schemas/UserType'
        status:
          $ref: '#/components/schemas/UserStatus'
        lastLoginTime:
          type: string
          format: date-time
        createTime:
          type: string
          format: date-time
    AuditRequest:
      type: object
      required:
      - auditStatus
      properties:
        auditStatus:
          type: string
          enum:
          - APPROVED
          - REJECTED
        auditReason:
          type: string
...
    OperationLog:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        module:
          type: string
        operation:
          type: string
        requestMethod:
          type: string
        requestUrl:
          type: string
        requestIp:
          type: string
        resultStatus:
          type: string
        costTime:
          type: integer
          format: int64
        createTime:
          type: string
          format: date-time
    LoginLog:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        userType:
          type: string
        loginIp:
          type: string
        status:
          type: string
        failReason:
          type: string
        loginTime:
          type: string
          format: date-time
...
    ApplicationPage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/Application'
        total:
          type: integer
          format: int64
          example: 100
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
...
    AdminUserPage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        total:
          type: integer
          format: int64
          example: 100
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
...
    OperationLogPage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/OperationLog'
        total:
          type: integer
          format: int64
          example: 100
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
    LoginLogPage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/LoginLog'
        total:
          type: integer
          format: int64
          example: 100
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
...
    ApiResponseApplication:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/Application'
...
    ApiResponseApplicationPage:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/ApplicationPage'
...
    ApiResponseAdminDashboard:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/AdminDashboard'
    ApiResponseAdminUserPage:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/AdminUserPage'
...
    ApiResponseOperationLogPage:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/OperationLogPage'
    ApiResponseLoginLogPage:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/LoginLogPage'
...